requirements.txt
Flask
gunicorn
openai-whisper
edge-tts
pydub
ffmpeg-python
numpy
soundfile
librosa
openai-whisper
scipy